Fix: Correct CurrencyPriceChartData model Add: Include ChartViewModel, CurrencyDetailsPageViewModel Fix: Resolve issues in CurrencyDetailsViewModel and CurrencyPriceChartDataViewModel

+++ src/Core/CWA.Domain/Enteties/CurrencyPriceChartData.cs
@@ -2,6 +2,6 @@
 {
     public class CurrencyPriceChartData
     {
-        public List<List<decimal>> Prices { get; set; }
+        public List<List<long>> Prices { get; set; }
     }
 }
